+++ src/views/login.html
@@ -7,8 +7,7 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Sources</title>
 
-  <!-- Google Font -->
-  <link href="https://fonts.googleapis.com/css?family=Roboto:400,500" rel="stylesheet">
+  <!-- Google Fonts -->
 
   <!-- Our Styles -->
   <link rel="stylesheet" href="./login.css">
